requirements.txt
streamlit
newspaper3k
openai
python-dotenv
lxml_html_clean
